backend/requirements.txt
Flask==2.2.1
Flask-Cors==3.0.10
mercadopago==2.2.1
Flask-RESTful==0.3.9
python-dotenv==1.0.0
Flask-HTTPAuth==4.5.0
Flask-SQLAlchemy==3.0.3